refactor and introduce package de.metas.inoutcandidate.picking_bom
https://github.com/metasfresh/metasfresh/issues/5679
