+++ de.metas.material/planning/src/main/java/de/metas/material/planning/IProductPlanningDAO.java
@@ -74,8 +74,4 @@ public interface IProductPlanningDAO extends ISingletonService
 	void setProductBOMIdIfAbsent(ProductId productId, ProductBOMId bomId);
 
 	Set<ProductBOMId> retrieveAllPickingBOMIds();
-
-	Optional<PickingOrderConfig> getPickingOrderConfig(OrgId orgId, WarehouseId warehouseId, ProductId productId, AttributeSetInstanceId asiId);
-
-	Optional<PickingOrderConfig> extractPickingOrderConfig(I_PP_Product_Planning productPlanning);
 }
