+++ de.metas.swat/de.metas.swat.base/src/main/java/de/metas/inoutcandidate/modelvalidator/PP_Product_Planning.java
@@ -1,34 +1,30 @@
-package org.eevolution.model.validator;
+package de.metas.inoutcandidate.modelvalidator;
 
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
 import org.adempiere.ad.modelvalidator.annotations.ModelChange;
-import org.adempiere.mm.attributes.AttributeSetInstanceId;
-import org.adempiere.mm.attributes.api.AttributesKeys;
-import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.model.ModelValidator;
 import org.eevolution.model.I_PP_Product_Planning;
+import org.springframework.stereotype.Component;
 
-import com.google.common.annotations.VisibleForTesting;
-
-import de.metas.material.event.commons.AttributesKey;
+import de.metas.inoutcandidate.picking_bom.PickingBOMService;
 import lombok.NonNull;
 
 /*
  * #%L
- * de.metas.adempiere.libero.libero
+ * de.metas.swat.base
  * %%
- * Copyright (C) 2017 metas GmbH
+ * Copyright (C) 2019 metas GmbH
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- *
+ * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- *
+ * 
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -36,22 +32,25 @@ import lombok.NonNull;
  */
 
 @Interceptor(I_PP_Product_Planning.class)
+@Component
 public class PP_Product_Planning
 {
+	private final PickingBOMService pickingBOMService;
+
+	public PP_Product_Planning(@NonNull final PickingBOMService pickingBOMService)
+	{
+		this.pickingBOMService = pickingBOMService;
+	}
+
 	@ModelChange(timings = { ModelValidator.TYPE_BEFORE_NEW, ModelValidator.TYPE_BEFORE_CHANGE })
 	public void beforeSave(@NonNull final I_PP_Product_Planning productPlanning)
 	{
-		if (InterfaceWrapperHelper.isValueChanged(productPlanning, I_PP_Product_Planning.COLUMNNAME_M_AttributeSetInstance_ID))
+		//
+		// Make sure the picking order configuration, if any, is valid
+		if (productPlanning.isPickingOrder())
 		{
-			updateStorageAttributesKey(productPlanning);
+			pickingBOMService.extractPickingOrderConfig(productPlanning); // extract it just to validate
 		}
 	}
 
-	@VisibleForTesting
-	void updateStorageAttributesKey(@NonNull final I_PP_Product_Planning productPlanning)
-	{
-		final AttributeSetInstanceId asiId = AttributeSetInstanceId.ofRepoIdOrNone(productPlanning.getM_AttributeSetInstance_ID());
-		final AttributesKey attributesKey = AttributesKeys.createAttributesKeyFromASIStorageAttributes(asiId).orElse(AttributesKey.NONE);
-		productPlanning.setStorageAttributesKey(attributesKey.getAsString());
-	}
 }
